# microservice/gateway/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "The gateway of forum",
        "title": "forum-gateway",
        "contact": {},
        "version": "1.0"
    },
    "host": "forum.muxixyz.com",
    "basePath": "/api/v1",
    "paths": {
...
        "/auth/login/team": {
            "post": {
                "description": "login the team-forum",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "团队登录 api",
                "parameters": [
                    {
                        "description": "login_request",
                        "name": "object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/TeamLoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/TeamLoginResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "TeamLoginRequest": {
            "type": "object",
            "properties": {
                "oauth_code": {
                    "type": "string"
                }
            }
        },
        "TeamLoginResponse": {
            "type": "object",
            "properties": {
                "redirect_url": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                }
            }
        },
...
    },
...
}